requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
click==8.1.7
et-xmlfile==1.1.0
fake-useragent==1.5.1
Flask==3.0.3
Flask-Session==0.8.0
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
magic-filter==1.0.12
MarkupSafe==2.1.5
msgspec==0.18.6
multidict==6.0.5
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
XlsxWriter==3.2.0
yarl==1.9.4
